requirements.txt
pandas
gradio
python-dotenv
openai